// src/app/core/api/product/product.ts
export interface Product {
    id?: string;
    code?: string;

    name?: string;
    email?: string;
    tel?: string,
    parkingSpot?: string,
    slotNumber?: string,
    assignmentDate?: string,
    date?: string ,
    status?: string,
    discountNote?: string,
    actions?: string,

    description?: string;
    price?: number;
    quantity?: number;
    inventoryStatus?: string;
    category?: string;
    image?: string;
    rating?: number;
}